reports/19073847305/40/requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.40.0
allure-pytest==2.13.2
python-dotenv==1.0.0
allure-python-commons~=2.13.2